requirements.txt
django
djangorestframework
markdown
django-filter
django-widget-tweaks
djangorestframework_simplejwt